% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/circleD3.r
\name{circleD3-shiny}
\alias{circleD3-shiny}
\alias{circleD3Output}
\alias{renderCircleD3}
\title{Shiny bindings for circleD3}
\usage{
circleD3Output(outputId, width = "100\%", height = "400px")

renderCircleD3(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{output variable to read from}

\item{width,height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{expr}{An expression that generates a circleD3}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
  is useful if you want to save an expression in a variable.}
}
\description{
Output and render functions for using circleD3 within Shiny
applications and interactive Rmd documents.
}

